refactor: split fat setup support scenario tables (#4349)
Topic-split the six ≥336-line support scenario modules and enforce a 380-line per-file support ratchet so wave-7 Slice D cannot re-accumulate fat tables.
