+++ SPEC/program/repo-lint.md
@@ -117,7 +117,7 @@
 | `tool/check_colonizethis_setup_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_setup/lib/**` physical lines per file (wave-7 ≤300; shrink-only grandfather allowlist during transition; generated suffixes excluded); rule `repo.colonizethis_setup_lib_file_size` (Refs #4273, #4349) |
 | `tool/check_colonizethis_setup_test_mirrors_lib.dart` | Forbid loose Dart sources directly at the `packages/colonizethis_setup/test/` root: setup production code lives under `lib/src/setup/`, so the testing rule (`colonizethis-testing.mdc`) requires `test/` to mirror `lib/`. Every setup test (and its co-located support helper) must live in a mirroring subdirectory such as `test/setup/` (with optional `test/setup/plains/`, `test/setup/locked_assigner/` subtrees) rather than flat in `test/`. A file is flagged when the path remainder after the `packages/colonizethis_setup/test/` prefix contains no further separator; nested files such as `test/setup/validation_exceptions_test.dart` are compliant. Mirrors the sibling orders/combat layout gates; rule `repo.colonizethis_setup_test_mirrors_lib` (Refs #4273) |
 | `tool/check_colonizethis_setup_scenario_table_runner.dart` | Prefer-scenario-tables gate for `packages/colonizethis_setup/test/**`: flags long imperative `test('…') { … }` bodies (heuristic: opening `{` on the same line as `test(` with no surrounding `for (final scenario` within a short look-behind window) outside `setupPreferScenarioTablesAllowlist`. Wave-7 slice C (#4349) sets baseline allow-all **off**; remaining legitimately imperative suites stay on the explicit allowlist after densifying the ≥300-line suites into `support/*_scenarios.dart`; mirrors `repo.orders_test_prefer_scenario_tables`; rule `repo.colonizethis_setup_prefer_scenario_tables` (Refs #4273, #4349) |
-| `tool/check_colonizethis_setup_test_support_loc.dart` | Ratchet ceiling for all Dart under `packages/colonizethis_setup/test/setup/support/` (physical LOC; wave-7 slice C post-densify ceiling **4100**, measured ≈3995; wave-6 was **2700**; Slice D will split fat scenario modules and re-ratchet); fail on net regrowth beyond the ceiling; rule `repo.colonizethis_setup_test_support_loc` (Refs #4273, #4349) |
+| `tool/check_colonizethis_setup_test_support_loc.dart` | Ratchet for Dart under `packages/colonizethis_setup/test/setup/support/`: tree physical LOC ceiling **4100** (wave-7 slice D post-split measured 3996 + headroom; wave-6 was **2700**; slice C densify moved suite bodies into support so the tree total cannot return to 2700) and per-file ceiling **380** (fail at ≥380); fail on net tree regrowth or a fat scenario module; rule `repo.colonizethis_setup_test_support_loc` (Refs #4273, #4349) |
 | `tool/check_setup_test_use_shared_fixtures.dart` | Forbid re-inlined empty `Game(` / `WorldState(` shells and full `GameSetupConfig(` rebuilds under `packages/colonizethis_setup/test/**`; empty/minimal shells must use `package:colonizethis_test/game_test_fixtures.dart` `TestFixtures` (or a thin support delegate), and config variants must use `configWithOverrides` / `lockedFullInitConfig` in `init_game_orchestrator_test_support.dart` (that support file alone may construct `GameSetupConfig`); full-line comments ignored (Refs #4029). Additionally forbid running `runInitGame(config: X, options: defaultInitOptions)` with an identical config expression in two or more distinct test bodies of one file — such assert-only repeats must reuse the memoized `sharedInitGameResult(config)` harness; determinism double runs inside a single test body, calls with custom `InitGameOptions` or extra arguments (for example `generateRegion:`), and the support file itself are structurally exempt (Refs #4054); rule `repo.setup_test_use_shared_fixtures` |
 | `tool/check_orders_file_size.dart` | Enforce the two orders hot files `packages/colonizethis_orders/lib/src/orders/orders_application.dart` and `.../orders_application_completed_work.dart` at **1000 non-comment lines or fewer** each (fail when strictly greater), reusing `countNonCommentLinesFromSource`; the gate also fails when either named file is missing so it cannot silently rot if the orders module moves. Complements the repository-wide `repo.dart_file_non_comment_line_size` cap with an orders-specific message so the #3404 dedup/abstraction refactors keep these files bounded (Refs #3404); rule `repo.orders_file_size` |
 | `tool/check_orders_test_no_duplicate_fixtures.dart` | Forbid re-introducing a local copy of the shared game/world test fixtures anywhere under `packages/colonizethis_orders/test/**`: any file named `test_fixtures.dart` or any test file that redeclares a top-level `TestFixtures` class is a violation. Orders tests must import the canonical `package:colonizethis_test/game_test_fixtures.dart` factories (single source of truth) rather than duplicating them; references to the shared `TestFixtures` symbol are unaffected. Complements the testing-rule shared-fixtures policy (Refs #3714); rule `repo.orders_test_no_duplicate_fixtures` |
@@ -587,6 +587,9 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_ai_contracts/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4084).
 - Given the `colonizethis_setup` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.setup_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4086).
 - Given a `packages/colonizethis_setup/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.setup_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4086).
+- Given every `*.dart` file under `packages/colonizethis_setup/test/setup/support/` has fewer than 380 physical lines and the tree total is at or below 4100, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_setup_test_support_loc`, then the rule passes and exits `0` (Refs #4273, #4349).
+- Given a `packages/colonizethis_setup/test/setup/support/**` Dart file with 380 or more physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_setup_test_support_loc`, then the run fails naming that file and exits `1` (Refs #4349).
+- Given the support tree physical LOC exceeds 4100 while every file is under 380 lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_setup_test_support_loc`, then the run fails reporting the tree total and exits `1` (Refs #4273, #4349).
 - Given `app/lib/**` on `dev` where every `part` / `part of` directive is confined to the shrink-only grandfather allowlist in `tool/app_no_part_directives_grandfather.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4117).
 - Given an `app/lib/**` Dart file not on the grandfather allowlist that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4117).
 - Given a stale grandfather allowlist entry (missing file or file no longer containing a `part` directive), when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_part_directives`, then the run fails with a stale-entry message and exits `1` (Refs #4117).
