+++ packages/colonizethis_setup/test/setup/game_setup_creation_and_assignment_part1_segment1_test.dart
@@ -1,14 +1,15 @@
 // Ported from colonizethis_logic (Refs #4090 Slice C).
 import 'package:colonizethis_test/test.dart';
 
-import 'support/game_setup_creation_and_assignment_part1_segment1_scenarios.dart';
+import 'support/game_setup_creation_and_assignment_part1_segment1_create_scenarios.dart';
+import 'support/game_setup_creation_and_assignment_part1_segment1_repeat_scenarios.dart';
 import 'support/scenario_runner.dart';
 
 void main() {
   group('GameSetup', () {
-    runLabeledScenarios(
-      gameSetupCreationAndAssignmentPart1Segment1Scenarios(),
-      runRunnableScenario,
-    );
+    runLabeledScenarios([
+      ...gameSetupCreationAndAssignmentPart1Segment1CreateScenarios(),
+      ...gameSetupCreationAndAssignmentPart1Segment1RepeatScenarios(),
+    ], runRunnableScenario);
   });
 }
